owl_fmtext_curs_waddstr -> owl_fmtext_ow_addstr
Make the "put fmtext in a window" function take an owl_window * rather than
a WINDOW *, and propagate this change back out a few levels of callers,
notably factoring out owl_fmtext_curs_waddstr_without_search because that
context is carried in the owl_window.   This leads us to switching out
the WINDOW * in the owl_viewwin structure, and nuking the unused
owl_viewwin_set_curswin function.